@@ -4,14 +4,18 @@
   "description": "",
   "main": "index.js",
   "typings": "index.d.ts",
+  "type": "module",
   "scripts": {
     "build": "npm run clean && tsc && node index.js",
-    "clean": "rm -f index.js index.d.ts person.d.ts"
+    "clean": "rm -f index.js index.d.ts person.d.ts",
+    "postinstall": "patch-package"
   },
   "dependencies": {
-    "@types/node": "^13.13.4",
-    "typescript": "3.8.3",
-    "json-schema-to-typescript": "9.0.0"
+    "@kickstartds/json-schema-to-typescript": "^13.1.5",
+    "@types/node": "^17.0.33",
+    "patch-package": "^8.0.0",
+    "postinstall-postinstall": "^2.1.0",
+    "typescript": "^4.9.5"
   },
   "repository": {
     "type": "git",
