@@ -1,19 +1,28 @@
 {
-  "name": "json-schema-to-typescript",
-  "version": "13.1.1",
+  "name": "@kickstartds/json-schema-to-typescript",
+  "version": "13.1.7",
   "description": "compile json schema to typescript typings",
-  "main": "dist/src/index.js",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
   "bin": {
-    "json2ts": "dist/src/cli.js"
+    "json2ts": "dist/cli.js"
   },
-  "typings": "dist/src/index.d.ts",
+  "types": "./dist/index.d.ts",
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=16.0.0"
+  },
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
   },
   "scripts": {
     "build": "npm run lint && npm run clean && npm run build:browser && npm run build:server",
     "build:browser": "browserify src/index.ts -s jstt -p tsify > dist/bundle.js",
-    "build:server": "tsc -d",
+    "build:server": "tsup src/index.ts src/cli.ts --format cjs,esm --dts --clean",
+    "build:test": "tsup test/test.ts test/e2e --external ava",
     "clean": "shx rm -rf dist && mkdir dist",
     "format": "prettier \"{src,test}/*.ts\" --write",
     "format-check": "prettier \"{src,test}/*.ts\" --check",
@@ -21,9 +30,10 @@
     "tdd": "concurrently -r -p '' -k 'npm run watch' 'npm run watch:test'",
     "test": "npm run pre-test && ava --timeout=300s --verbose",
     "stresstest": "seq 1 10 | xargs -I{} npm test",
+    "postinstall": "patch-package",
     "prepublishOnly": "npm test",
-    "pre-test": "npm run clean && npm run format-check && npm run build:server",
-    "watch": "tsc -w",
+    "pre-test": "npm run clean && npm run format-check && npm run build:server && npm run build:test",
+    "watch": "tsup src/index.ts src/cli.ts --watch",
     "watch:test": "ava -w"
   },
   "repository": {
@@ -41,6 +51,7 @@
     "typing",
     "share"
   ],
+  "type": "module",
   "author": "Boris Cherny <boris@performancejs.com> (http://performancejs.com/)",
   "license": "MIT",
   "bugs": {
@@ -50,31 +61,32 @@
   "dependencies": {
     "@bcherny/json-schema-ref-parser": "10.0.5-fork",
     "@types/json-schema": "^7.0.11",
-    "@types/lodash": "^4.14.182",
+    "@types/lodash-es": "^4.17.9",
     "@types/prettier": "^2.6.1",
     "cli-color": "^2.0.2",
     "get-stdin": "^8.0.0",
     "glob": "^7.1.6",
     "glob-promise": "^4.2.2",
     "is-glob": "^4.0.3",
-    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
     "minimist": "^1.2.6",
     "mkdirp": "^1.0.4",
-    "mz": "^2.7.0",
+    "patch-package": "^8.0.0",
+    "postinstall-postinstall": "^2.1.0",
     "prettier": "^2.6.2"
   },
   "devDependencies": {
+    "@swc/core": "^1.3.82",
     "@types/cli-color": "^2.0.2",
     "@types/glob": "^7.2.0",
     "@types/is-glob": "^4.0.2",
     "@types/minimist": "^1.2.2",
     "@types/mkdirp": "^1.0.2",
-    "@types/mz": "^2.7.4",
     "@types/node": "^17.0.33",
     "@types/rimraf": "^3.0.2",
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
-    "ava": "^4.2.0",
+    "ava": "^5.3.1",
     "browserify": "^17.0.0",
     "browserify-shim": "^3.8.14",
     "concurrently": "^7.2.0",
@@ -84,16 +96,16 @@
     "rimraf": "^3.0.2",
     "shx": "^0.3.4",
     "tsify": "^5.0.4",
-    "typescript": "^4.6.4"
+    "tsup": "^7.2.0",
+    "typescript": "^4.9.5"
   },
   "ava": {
     "files": [
-      "./dist/test/test.js"
+      "./dist/test.cjs"
     ],
     "ignoredByWatcher": [
       "./src"
-    ],
-    "snapshotDir": "./test/__snapshots__"
+    ]
   },
   "browserify": {
     "transform": [
