@@ -1,6 +1,6 @@
-import $RefParser = require('@bcherny/json-schema-ref-parser')
-import {JSONSchema} from './types/JSONSchema'
-import {log} from './utils'
+import $RefParser from '@bcherny/json-schema-ref-parser'
+import {JSONSchema} from './types/JSONSchema.js'
+import {log} from './utils.js'
 
 export type DereferencedPaths = WeakMap<$RefParser.JSONSchemaObject, string>
 
