@@ -1,5 +1,5 @@
-import {writeFileSync} from 'fs'
-import {compileFromFile} from 'json-schema-to-typescript'
+import {writeFileSync} from 'node:fs'
+import {compileFromFile} from '@kickstartds/json-schema-to-typescript'
 
 async function generate() {
   writeFileSync('person.d.ts', await compileFromFile('person.json'))
