@@ -7,92 +7,90 @@ import rimraf = require('rimraf')
 export function run() {
   test('pipe in, pipe out', t => {
     t.snapshot(
-      execSync('shx cat ./test/resources/ReferencedType.json | node dist/src/cli.js', {encoding: 'utf-8'}).toString(),
+      execSync('shx cat ./test/resources/ReferencedType.json | node dist/cli.cjs', {encoding: 'utf-8'}).toString(),
     )
   })
 
   test('pipe in (schema without ID), pipe out', t => {
     t.snapshot(
-      execSync('shx cat ./test/resources/ReferencedTypeWithoutID.json | node dist/src/cli.js', {
+      execSync('shx cat ./test/resources/ReferencedTypeWithoutID.json | node dist/cli.cjs', {
         encoding: 'utf-8',
       }).toString(),
     )
   })
 
   test('file in (no flags), pipe out', t => {
-    t.snapshot(execSync('node dist/src/cli.js ./test/resources/ReferencedType.json').toString())
+    t.snapshot(execSync('node dist/cli.cjs ./test/resources/ReferencedType.json').toString())
   })
 
   test('file in (--input), pipe out', t => {
-    t.snapshot(execSync('node dist/src/cli.js --input ./test/resources/ReferencedType.json').toString())
+    t.snapshot(execSync('node dist/cli.cjs --input ./test/resources/ReferencedType.json').toString())
   })
 
   test('file in (-i), pipe out', t => {
-    t.snapshot(execSync('node dist/src/cli.js -i ./test/resources/ReferencedType.json').toString())
+    t.snapshot(execSync('node dist/cli.cjs -i ./test/resources/ReferencedType.json').toString())
   })
 
   test('file in (-i), unreachable definitions flag, pipe out', t => {
     t.snapshot(
-      execSync('node dist/src/cli.js -i ./test/resources/DefinitionsOnly.json --unreachableDefinitions').toString(),
+      execSync('node dist/cli.cjs -i ./test/resources/DefinitionsOnly.json --unreachableDefinitions').toString(),
     )
   })
 
   test('file in (-i), style flags, pipe out', t => {
     t.snapshot(
-      execSync('node dist/src/cli.js -i ./test/resources/Enum.json --style.singleQuote --no-style.semi').toString(),
+      execSync('node dist/cli.cjs -i ./test/resources/Enum.json --style.singleQuote --no-style.semi').toString(),
     )
   })
 
   test('file in (-i), pipe out (absolute path)', t => {
-    t.snapshot(execSync(`node dist/src/cli.js -i ${__dirname}/../../test/resources/ReferencedType.json`).toString())
+    t.snapshot(execSync(`node dist/cli.cjs -i ${__dirname}/../test/resources/ReferencedType.json`).toString())
   })
 
   test('pipe in, file out (--output)', t => {
-    execSync('shx cat ./test/resources/ReferencedType.json | node dist/src/cli.js --output ./ReferencedType.d.ts')
+    execSync('shx cat ./test/resources/ReferencedType.json | node dist/cli.cjs --output ./ReferencedType.d.ts')
     t.snapshot(readFileSync('./ReferencedType.d.ts', 'utf-8'))
     unlinkSync('./ReferencedType.d.ts')
   })
 
   test('pipe in, file out (-o)', t => {
-    execSync('shx cat ./test/resources/ReferencedType.json | node dist/src/cli.js -o ./ReferencedType.d.ts')
+    execSync('shx cat ./test/resources/ReferencedType.json | node dist/cli.cjs -o ./ReferencedType.d.ts')
     t.snapshot(readFileSync('./ReferencedType.d.ts', 'utf-8'))
     unlinkSync('./ReferencedType.d.ts')
   })
 
   test('file in (no flags), file out (no flags)', t => {
-    execSync('node dist/src/cli.js ./test/resources/ReferencedType.json ./ReferencedType.d.ts')
+    execSync('node dist/cli.cjs ./test/resources/ReferencedType.json ./ReferencedType.d.ts')
     t.snapshot(readFileSync('./ReferencedType.d.ts', 'utf-8'))
     unlinkSync('./ReferencedType.d.ts')
   })
 
   test('file in (-i), file out (-o)', t => {
-    execSync('node dist/src/cli.js -i ./test/resources/ReferencedType.json -o ./ReferencedType.d.ts')
+    execSync('node dist/cli.cjs -i ./test/resources/ReferencedType.json -o ./ReferencedType.d.ts')
     t.snapshot(readFileSync('./ReferencedType.d.ts', 'utf-8'))
     unlinkSync('./ReferencedType.d.ts')
   })
 
   test('file in (--input), file out (--output)', t => {
-    execSync('node dist/src/cli.js --input ./test/resources/ReferencedType.json --output ./ReferencedType.d.ts')
+    execSync('node dist/cli.cjs --input ./test/resources/ReferencedType.json --output ./ReferencedType.d.ts')
     t.snapshot(readFileSync('./ReferencedType.d.ts', 'utf-8'))
     unlinkSync('./ReferencedType.d.ts')
   })
 
   test('--unknownAny', t => {
-    t.snapshot(
-      execSync('node dist/src/cli.js --unknownAny=false --input ./test/resources/ReferencedType.json').toString(),
-    )
+    t.snapshot(execSync('node dist/cli.cjs --unknownAny=false --input ./test/resources/ReferencedType.json').toString())
   })
 
   test('--additionalProperties', t => {
     t.snapshot(
       execSync(
-        'node dist/src/cli.js --additionalProperties=false --input ./test/resources/ReferencedType.json',
+        'node dist/cli.cjs --additionalProperties=false --input ./test/resources/ReferencedType.json',
       ).toString(),
     )
   })
 
   test('files in (-i), files out (-o)', t => {
-    execSync(`node dist/src/cli.js -i "./test/resources/MultiSchema/**/*.json" -o ./test/resources/MultiSchema/out`)
+    execSync(`node dist/cli.cjs -i "./test/resources/MultiSchema/**/*.json" -o ./test/resources/MultiSchema/out`)
 
     readdirSync('./test/resources/MultiSchema/out').forEach(f => {
       const path = `./test/resources/MultiSchema/out/${f}`
@@ -104,12 +102,12 @@ export function run() {
   })
 
   test('files in (-i), pipe out', t => {
-    t.snapshot(execSync(`node dist/src/cli.js -i "./test/resources/MultiSchema/**/*.json"`).toString())
+    t.snapshot(execSync(`node dist/cli.cjs -i "./test/resources/MultiSchema/**/*.json"`).toString())
   })
 
   test('files in (-i), files out (-o) nested dir does not exist', t => {
     execSync(
-      `node dist/src/cli.js -i "./test/resources/MultiSchema/**/*.json" -o ./test/resources/MultiSchema/foo/bar/out`,
+      `node dist/cli.cjs -i "./test/resources/MultiSchema/**/*.json" -o ./test/resources/MultiSchema/foo/bar/out`,
     )
     readdirSync('./test/resources/MultiSchema/foo/bar/out').forEach(f => {
       const path = `./test/resources/MultiSchema/foo/bar/out/${f}`
@@ -122,7 +120,7 @@ export function run() {
 
   test('files in (-i), files out (-o) matching nested dir', t => {
     execSync(
-      `node dist/src/cli.js -i "./test/resources/../../test/resources/MultiSchema2/" -o ./test/resources/MultiSchema2/out`,
+      `node dist/cli.cjs -i "./test/resources/../../test/resources/MultiSchema2/" -o ./test/resources/MultiSchema2/out`,
     )
     getPaths('./test/resources/MultiSchema2/out').forEach(file => {
       t.snapshot(file)
